Refactor of lazy services
Lazy services were added after the ServiceManager had an established API, and
as such had to rely on optional integration. This patch alters that relationship
to make lazy service configuration a first-class citizen of the ServiceManager.

This allows:

- Removing the `LazyServiceFactoryFactory`.
- Passing `lazy_services` configuration to the constructor or `withConfig()`
  method.
- Direct instantiation of the `LazyServiceFactory` when first encountered as a
  delegator factory, using the `lazy_services` configuration composed in the
  ServiceManager itself.

The usage is thus simplified, as all configuration for all service types is now
directly handled by the ServiceManager itself.
