+++ doc/book/lazy-services.md
@@ -3,7 +3,7 @@
 `Zend\ServiceManager` can use [delegator factories](delegators.md) to generate
 "lazy" references to your services.
 
-Lazy services are [proxies](http://en.wikipedia.org/wiki/Proxypattern)_ that
+Lazy services are [proxies](http://en.wikipedia.org/wiki/Proxypattern) that
 get lazily instantiated, and keep a reference to the real instance of
 the proxied service.
 
@@ -24,10 +24,8 @@ object is really needed.
 
 ## Setup
 
-`Zend\ServiceManager\Factory\LazyServiceFactoryFactory` is a
-[factory](configuring-the-service-manager.md#factories) capable of generating
-lazy loading proxies for your services. It returns a `Zend\Service\Proxy\LazyServiceFactory`,
-which will return a proxy instance for the service requested.
+`Zend\ServiceManager\Proxy\LazyServiceFactory` is a [delegator factory](delegators.md)
+capable of generating lazy loading proxies for your services.
 
 The lazy service facilities depend on [ProxyManager](https://github.com/Ocramius/ProxyManager);
 you will need to install that package before using the feature:
@@ -66,27 +64,19 @@ instead of real services:
 ```php
 use MyApp\Buzzer;
 use Zend\ServiceManager\Factory\InvokableFactory;
-use Zend\ServiceManager\Factory\LazyServiceFactoryFactory;
 use Zend\ServiceManager\Proxy\LazyServiceFactory;
 use Zend\ServiceManager\ServiceManager;
 
-$config = [
+$serviceManager = new \Zend\ServiceManager\ServiceManager([
+    'factories' => [
+        Buzzer::class             => InvokableFactory::class,
+    ],
     'lazy_services' => [
-         // mapping services to their class names is required
-         // since the ServiceManager is not a declarative DIC
+         // Mapping services to their class names is required
+         // since the ServiceManager is not a declarative DIC.
          'class_map' => [
              Buzzer::class => Buzzer::class,
          ],
-    ],
-];
-
-$serviceManager = new \Zend\ServiceManager\ServiceManager([
-    'services' => [
-        'config' => $config,
-    ],
-    'factories' => [
-        Buzzer::class             => InvokableFactory::class,
-        LazyServiceFactory::class => LazyServiceFactoryFactory::class,
     ],
     'delegators' => [
         Buzzer::class => [
@@ -126,7 +116,7 @@ echo $buzzer->buzz();
 ## Configuration
 
 This is the config structure expected by `Zend\ServiceManager\Proxy\LazyServiceFactory`,
-in the `lazy_services` key passed in the application configuration:
+in the `lazy_services` key passed in the service manager configuration:
 
 ```php
 [
